Speed up pretty svg with a memo table
This change introduces a helper that lets `pretty-src-svg` use a git note to avoid rebuilding the same output for the same input on every run.  On my machine, this reduces the build time from minutes to seconds for incremental updates.  This would result in a similar speed-up between CI and CD runs where the git caches are preserved between runs.