Remove use of default colours from emoji.
This caused unexpected display differences. A common symptom was eyes that matched the text colour, this resulted in poor results when the text colour isn't black. This sets the fill and stroke colour to black for any elements where they weren't specified and it was likely expected.

This change was generated by running the following commands to add stroke and fill attributes to elements which didn't have them:

```sh
cd src/

comby -jobs 32 '<:[e~ellipse|circle|rect] :[r]>' '<:[e] fill="#000" :[r]>' .svg -in-place
comby -jobs 32 '<path fill="#000" :[a]stroke=:[b]>' '<path :[a]stroke=:[b]>' .svg -in-place
comby -jobs 32 '<:[e] fill="#000" :[a]fill=:[b]>' '<:[e] :[a]fill=:[b]>' .svg -in-place
```

(This script works by adding the attributes to all instances of the intended elements then removing it if it causes a duplicate.)

The results were then manually reviewed to revert emoji where it makes sense to match the text colour as well as fix up the emoji that were not handled correctly by the simple script.

Fixes a lot of https://github.com/hfg-gmuend/openmoji/issues/339