Remove duplicate id attributes from certain SVG elements.
Some of the SVG files have multiple elements with the same id.
This violates the specifications for the SVG format, and caused problems when I was trying to parse the files with another program.

Comparing these files to the other SVG sources for OpenMoji, I found that these id attributes are nonstandard.
Most of the SVG files have descriptive ids for each group ("line", "color", "skin", etc.), but not for the elements within those groups.
As such, the easiest fix was to simply remove those problematic id attributes.