Convert several transform attributes to matrix notation.
Strictly speaking, this isn't a fix to OpenMoji's src files.
Rather, this is a workaround.

When the OpenMoji-Black font is viewed in Chrome or Safari on MacOS, the height of lines is far too high. This is because several glyphs are rendered with small fragments far away from where they should be.

The glyphs broken in this way are: 🔬 🪖 🪄 🪘 🫔   and .

The commonality between these glyphs is that their source svg features an element in the *line* or *line-supplement* group with a rotation about a point other than the origin.
When the resulting font is viewed, some browsers instead view the font as if that element were rotated about (0,0).
In most cases, this results in the element being placed far outside the normal bounding box of the character.

The same problem in the same glyphs arose when I tried to build an OpenMoji font myself using FontForge.
If OpenMoji's current build process also uses FontForge at some stage, then I suspect the error might ultimately lay somewhere in that software's import function.

I'm hesitant to suggest a change to the source files for OpenMoji to deal with bugs in another piece of software.
However, the changes I have made are still vanilla SVG, and don't fundamentally alter the content of the files.

I identified 8 files with this problem, but this commit only altered 6.
This is because for two of these files (the Chrome and Chromium logos), the transformation didn't actually *do* anything, and so removing the attribute entirely was an unambiguous improvement.