Minor modifications to improve font contours.
Each of these modifications should have negligible or no effects on the appearance of the svg, while improving the consistency of the source files. The changes fall into the following main categories:
- Explicitly closing paths which loop around to their original location without closing.
- Removing invisible elements from the line layer.
- Removing elements which were effectively duplicates (paths which were subsets of other paths, things like that.)
- Applying transforms and collapsing groups within the line layer.

There were a few instances where I changed the direction of a path to manually fix a FontForge fill error. (eg see 1F532.svg)