Remove superfluous points in paths of 🎩
The outline of the hat had some colinear points that didn't affect the appearance of the curve at all. The fill of the top of the hat had those same points and also the colinear points lower down on the hat. For an object without a stroke, those aren't even visible.